Cast integers to ints before formatting in NSString

+++ DAAppsViewController/DAAppsViewController.m
@@ -170,7 +170,7 @@
 
 - (void)loadAppsWithArtistId:(NSInteger)artistId completionBlock:(void(^)(BOOL result, NSError *error))block
 {
-    NSString *requestPath = [NSString stringWithFormat:@"lookup?id=%lu", artistId];
+    NSString *requestPath = [NSString stringWithFormat:@"lookup?id=%i", (int)artistId];
     [self loadAppsWithPath:requestPath defaultTitle:NSLocalizedString(@"Results",) completionBlock:block];
 }
 
@@ -257,7 +257,7 @@
     }
     
     if ([SKStoreProductViewController class]) {
-        NSString *itunesItemIdentifier = [NSString stringWithFormat:@"%lu", appObject.appId];
+        NSString *itunesItemIdentifier = [NSString stringWithFormat:@"%i", (int)appObject.appId];
         NSDictionary *appParameters = @{SKStoreProductParameterITunesItemIdentifier: itunesItemIdentifier};
         SKStoreProductViewController *productViewController = [[SKStoreProductViewController alloc] init];
         [productViewController setDelegate:self];
@@ -266,7 +266,7 @@
                            animated:YES
                          completion:nil];
     } else {
-        NSString *appUrlString = [NSString stringWithFormat:@"itms-apps://itunes.apple.com/app/id%lu?mt=8", appObject.appId];
+        NSString *appUrlString = [NSString stringWithFormat:@"itms-apps://itunes.apple.com/app/id%i?mt=8", (int)appObject.appId];
         NSURL *appURL = [[NSURL alloc] initWithString:appUrlString];
         [[UIApplication sharedApplication] openURL:appURL];
     }
